+++ Front/src/app/dashboard/adminDashboard/page.tsx
@@ -1,7 +1,8 @@
 "use client";
 import React, { useState } from "react";
-import DashboardAdminCompo from "@/components/DashboardAdminCompo/DashboardAdminCompo";
 import UserList from "@/components/UserList/UserList";
+import Services from "@/components/Services/Services";
+import ListGardeners from "@/components/ListGardeners/ListGardeners";
 
 const AdminDashboard = () => {
   const [activeComponent, setActiveComponent] = useState<string>(""); // Controla el componente activo
@@ -14,23 +15,32 @@ const AdminDashboard = () => {
         <div className="flex gap-4">
           <button
             onClick={() => setActiveComponent("userList")}
-            className={`px-4 py-2 rounded-lg transition ${
-              activeComponent === "userList"
+            className={`px-4 py-2 rounded-lg transition ${activeComponent === "userList"
                 ? "bg-white text-green-600"
                 : "bg-green-700 hover:bg-green-800"
-            }`}
+              }`}
           >
             Lista de Usuarios
           </button>
           <button
             onClick={() => setActiveComponent("services")}
-            className={`px-4 py-2 rounded-lg transition ${
-              activeComponent === "services"
+            className={`px-4 py-2 rounded-lg transition ${activeComponent === "services"
                 ? "bg-white text-green-600"
                 : "bg-green-700 hover:bg-green-800"
-            }`}
+              }`}
           >
             Servicios Disponibles
+          </button>
+
+
+          <button
+            onClick={() => setActiveComponent("ListGardeners")}
+            className={`px-4 py-2 rounded-lg transition ${activeComponent === "ListGardeners"
+                ? "bg-white text-green-600"
+                : "bg-green-700 hover:bg-green-800"
+              }`}
+          >
+            Lista de Jardineros
           </button>
         </div>
       </div>
@@ -44,10 +54,14 @@ const AdminDashboard = () => {
         )}
         {activeComponent === "services" && (
           <div className="bg-white p-6 rounded-lg shadow-md">
-            <DashboardAdminCompo />
+            <Services />
           </div>
         )}
-
+        {activeComponent === "ListGardeners" && (
+          <div className="bg-white p-6 rounded-lg shadow-md">
+            <ListGardeners/>
+          </div>
+        )}
         {/* Mensaje inicial si no hay selección */}
         {!activeComponent && (
           <div className="text-gray-500 text-center mt-20">
