+++ Front/src/components/UserList/UserList.tsx
@@ -1,7 +1,6 @@
-"use client";
-
 import React, { useState, useEffect } from "react";
-import { banUser, getAllUsers } from "@/helpers/userOrders.helpers";
+import { banUser, deleteUser, getAllUsers } from "@/helpers/userOrders.helpers";
+import { log } from "console";
 
 interface User {
   id: string;
@@ -39,7 +38,7 @@ function UserList() {
     try {
       const token = localStorage.getItem("userSession");
       if (token) {
-        await banUser(userId, token, isBanned);
+        await banUser(userId, isBanned);
 
         const updatedUsers = await getAllUsers();
         setUsers(updatedUsers.data);
@@ -54,6 +53,27 @@ function UserList() {
       setBanError(error.message || "Error al banear/desbanear usuario");
     } finally {
       setIsBanning(false);
+    }
+  };
+
+  const handleDeleteUser = async (userId: string) => {
+    const confirmDelete = window.confirm("¿Estás seguro de que quieres eliminar este usuario?");
+    if (!confirmDelete) return;
+
+    try {
+      const token = localStorage.getItem("userSession");
+      if (token) {
+        await deleteUser(userId);
+        const updatedUsers = await getAllUsers();
+        setUsers(updatedUsers.data);
+        alert("Usuario eliminado correctamente");
+      } else {
+        console.error("User token not found");
+        setBanError("Token de usuario no encontrado");
+      }
+    } catch (error: any) {
+      console.error("Error deleting user:", error);
+      alert("Error al eliminar el usuario");
     }
   };
 
@@ -82,16 +102,25 @@ function UserList() {
               <p className="text-gray-600">Estado: {user.isBanned ? "Baneado" : "Activo"}</p>
               <p className="text-gray-600">Rol: {user.role}</p>
 
+              {/* Botón de banear/desbanear */}
               <button
                 onClick={() => handleBanUnbanUser(user.id, user.isBanned)}
-                className={`mt-4 px-4 py-2 rounded-md font-medium transition ${
-                  user.isBanned
+                className={`mt-4 px-4 py-2 rounded-md font-medium transition ${user.isBanned
                     ? "bg-red-500 text-white hover:bg-red-600"
                     : "bg-green-500 text-white hover:bg-green-600"
-                } disabled:bg-gray-400 disabled:cursor-not-allowed`}
+                  } disabled:bg-gray-400 disabled:cursor-not-allowed`}
                 disabled={isBanning}
               >
                 {isBanning ? "Procesando..." : user.isBanned ? "Desbanear" : "Banear"}
+              </button>
+
+              {/* Botón de eliminar */}
+              <button
+                onClick={() => handleDeleteUser(user.id)}
+                className="mt-4 px-4 py-2 rounded-md font-medium bg-red-700 text-white hover:bg-red-800 transition disabled:bg-gray-400 disabled:cursor-not-allowed"
+                disabled={isBanning}
+              >
+                Eliminar
               </button>
             </div>
           ))}
