+++ Front/src/helpers/service.helpers.ts
@@ -24,3 +24,4 @@ export const getServicesProvided = async(): Promise<IService[]> => {
     return [];
   }
 };
+
